When copying/moving,, add message that shows if file contents the same. Currently hard coded to only do file contents comparing if both files under 10MB.
The 10MB restriction was chosen because large files (say > 100MB) can take seconds to the comparison checks, which will make the UI appear sluggish to the user.

Doing straight memcmp with buffers to limit memory and for comparison speed.  md5 or other hashing checks are too computationally intensive, and the comparison is one-time (don't need to remember hash values)

Also, removed another printf.